refactor(managers): moved AIContextManager to Infrastructure.Managers
